Move types into separate files
